Added option in CPBarPlot to set bar widths and offsets in either view coordinates, or in plot coordinates. Plot coordinates are the default.
In order to handle both cases, a BOOL property (barWidthsAreInViewCoordinates) has been included. Also, the barWidth and barOffset properties are now NSDecimals, which is required if plot coordinates are used.

The previous implementation was somewhat flawed, using insufficient precision when in plot coordinates. This has been fixed in this commit.